remove unnecessary catch, migrate loader into single file where it is used
